Move image convolution to more logical place
