+++ analysis/datasets/voyage_hdf5_dataset.py
@@ -2,7 +2,7 @@ import h5py
 import torch
 
 from astromorph.astromorph.src.datasets.base_dataset import BaseDataset
-from make_thicker_line_plots import convolve_image
+from .helpers import convolve_image
 
 
 class VoyageHDF5Dataset(BaseDataset):
